Changes in Asset model in order to fix AssetsIntegrationTest.pushAsset() and SearchIntegrationTest. We have to check if floating point value can serve as a valid sequence number and revert this commit if it can't

+++ sdk/src/main/java/com/ambrosus/sdk/Asset.java
@@ -49,7 +49,7 @@ public class Asset extends Entity {
         return content.idData.getTimestamp();
     }
 
-    public long getSequenceNumber() {
+    public double getSequenceNumber() {
         return content.idData.sequenceNumber;
     }
 
@@ -74,7 +74,7 @@ public class Asset extends Entity {
 
     static class AssetIdData extends IdData {
 
-        private long sequenceNumber;
+        private double sequenceNumber;
 
         //no-argument contructor for GSON
         private AssetIdData(){
